requirements.txt
# AI Tech Bytes - Python Dependencies
# Install with: pip install -r requirements.txt

# News fetching
requests>=2.31.0
feedparser>=6.0.0

# Text-to-Speech
gTTS>=2.5.0

# Text Summarization (AI)
transformers>=4.30.0
torch>=2.0.0

# Video processing
moviepy>=1.0.3
imageio>=2.31.1
imageio-ffmpeg>=0.4.9

# Image processing (for moviepy)
Pillow>=10.0.0
numpy>=1.24.0

# YouTube upload (optional - requires OAuth setup)
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Utility
python-dotenv>=1.0.0  # For environment variables
